+++ css/moviesStyle.css
@@ -1,30 +1,30 @@
 * {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
 }
 
-body{
-    background-color: white;
+body {
+  background-color: white;
 }
-img{
-    width: 100%;
-    height: 30vh;
-    object-fit: cover; 
-    border-radius: 40px 40px 0 0;
+img {
+  width: 100%;
+  height: 30vh;
+  object-fit: cover;
+  border-radius: 40px 40px 0 0;
 }
 
-.heading-container{
-    background-color: blue;
-    color: white;
-    font-size: 2.5;
-    padding: 0.8em 0;
-    position: fixed;
-    top: 0;
-    width: 100%;
-    text-align: center;
-    box-shadow: 0 10px 50px 0 black; 
-    z-index:1;
+.heading-container {
+  background-color: blue;
+  color: white;
+  font-size: 2.5;
+  padding: 0.8em 0;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  text-align: center;
+  box-shadow: 0 10px 50px 0 black;
+  z-index: 1;
 }
 
 .heading-container h4 {
@@ -33,23 +33,23 @@ img{
   letter-spacing: 5px;
 }
 
-.options-container{
-    margin: 10em 0 2em 0;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 1em;
-    text-align: center;
+.options-container {
+  margin: 10em 0 2em 0;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 1em;
+  text-align: center;
 }
 
-.option{
-    background-color: white;
-    border: none;
-    cursor: pointer;
-    padding: 0.8em 1.5em;
-    border-radius: 0.5em;
-    border: none;
-    text-transform: capitalize;
+.option {
+  background-color: white;
+  border: none;
+  cursor: pointer;
+  padding: 0.8em 1.5em;
+  border-radius: 0.5em;
+  border: none;
+  text-transform: capitalize;
 }
 
 .container {
@@ -59,49 +59,48 @@ img{
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   grid-column-gap: 1em;
-
 }
 
-.movies-card{
-    display: grid;
-    position: relative;
-    grid-template-rows: auto 1fr;
-    box-shadow: 0 0 15px gray;
-    margin: 1em 0;
-    border-radius: 40px 40px;
+.movies-card {
+  display: grid;
+  position: relative;
+  grid-template-rows: auto 1fr;
+  box-shadow: 0 0 15px gray;
+  margin: 1em 0;
+  border-radius: 40px 40px;
 }
 
 .movies-card-img {
-    border-radius: 40px
+  border-radius: 40px;
 }
 
 .movies-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    padding: 1em;
-    font-size: 15px;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 1em;
+  font-size: 15px;
+  align-items: center;
 }
 
 .movies-title {
-    font-weight: 600;
-    color: black;
+  font-weight: 600;
+  color: black;
 }
 
 .movies-description {
-    color: gray;
-    text-align: justify;
-    margin: 0.8em 0 1ed 0;
+  color: gray;
+  text-align: justify;
+  margin: 0.8em 0 1ed 0;
 }
 
 .view-button {
-    position: relative;
-    text-decoration: none;
-    background-color: blue;
-    color: white;
-    width: 7em;
-    text-align: center;
-    padding: 0.3em 0.2em;
-    border-radius: 10px;
+  position: relative;
+  text-decoration: none;
+  background-color: blue;
+  color: white;
+  width: 7em;
+  text-align: center;
+  padding: 0.3em 0.2em;
+  border-radius: 10px;
 }
