+++ css/playerCard.css
@@ -3,92 +3,65 @@
   padding: 0;
   box-sizing: border-box;
 }
-section {
-  position: fixed;
-  height: 100%;
-  width: 100%;
-  background: #e3f2fd;
-}
-button {
-  font-size: 18px;
-  font-weight: 400;
-  color: #fff;
-  padding: 14px 22px;
-  border: none;
-  background: #4070f4;
-  border-radius: 6px;
-  cursor: pointer;
-}
-button:hover {
-  background-color: #265df2;
-}
-button.show-modal,
-.modal-box {
-  position: fixed;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  box-shadow: 0 5px 10px rgba(0, 0, 0, 0.1);
-}
-section.active .show-modal {
-  display: none;
-}
-.overlay {
-  position: fixed;
-  height: 100%;
-  width: 100%;
-  background: rgba(0, 0, 0, 0.3);
-  opacity: 0;
-  pointer-events: none;
-}
-section.active .overlay {
-  opacity: 1;
-  pointer-events: auto;
-}
-.modal-box {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  max-width: 380px;
-  width: 100%;
-  padding: 30px 20px;
-  border-radius: 24px;
-  background-color: #fff;
-  opacity: 0;
-  pointer-events: none;
-  transition: all 0.3s ease;
-  transform: translate(-50%, -50%) scale(1.2);
-}
-section.active .modal-box {
-  opacity: 1;
-  pointer-events: auto;
-  transform: translate(-50%, -50%) scale(1);
-}
-.modal-box i {
-  font-size: 70px;
-  color: #4070f4;
-}
-.modal-box h2 {
-  margin-top: 20px;
-  font-size: 25px;
-  font-weight: 500;
-  color: #333;
-}
-.modal-box h3 {
-  font-size: 16px;
-  font-weight: 400;
-  color: #333;
-  text-align: center;
-}
-.modal-box .buttons {
-  margin-top: 25px;
-}
-.modal-box button {
-  font-size: 14px;
-  padding: 6px 12px;
-  margin: 0 10px;
+
+body {
+  background: black;
 }
 
-img {
-    scale: 0.1;
+.card-container {
+  display: block;
+  width: 50%;
+  margin-top: 5%;
+  margin-left: 25%;
+  background-color: lightblue;
+  justify-content: center;
+  text-align: center;
+  border-radius: 4%;
+  box-shadow: rgba(0, 0, 0, 0.17) 0px -23px 25px 0px inset,
+    rgba(0, 0, 0, 0.15) 0px -36px 30px 0px inset,
+    rgba(0, 0, 0, 0.1) 0px -79px 40px 0px inset, rgba(0, 0, 0, 0.06) 0px 2px 1px,
+    rgba(0, 0, 0, 0.09) 0px 4px 2px, rgba(0, 0, 0, 0.09) 0px 8px 4px,
+    rgba(0, 0, 0, 0.09) 0px 16px 8px, rgba(0, 0, 0, 0.09) 0px 32px 16px;
+}
+
+header {
+  padding-top: 50px;
+  border-radius: 45%;
+}
+
+h1 {
+  align-items: center;
+  color: rebeccapurple;
+}
+
+.cards {
+  display: flex;
+  justify-content: center;
+  padding-bottom: 7%;
+}
+
+.card {
+  margin-left: 10px;
+  margin-top: 50px;
+  background-image: linear-gradient(45deg, #4f639b, #fff);
+  border-style: solid;
+  border-color: black;
+}
+
+.card-title {
+  text-align: center;
+}
+
+.card-img-top {
+  width: 100%;
+  height: 100%;
+}
+
+.card-body {
+  display: flex;
+  flex-direction: column;
+}
+
+section {
+  display: block;
 }
