--- css/style.css
@@ -1,23 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    background-color: black;
-}
-
-#canvas {
-    border: 5px solid black;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%,-50%);
-    max-width: 100%;
-    max-height: 100%;
-}
-
-#player, #layer1,#layer2,#layer3,#layer4,#layer5,#layer6,#theSun, #enemy_fly,#enemy_plant, #enemy_spider_big, #fire, #collisionAnimation, #lives {
-    display: none;
-}
