+++ gamingNews.html
@@ -1,3 +1,5 @@
+<!-- Page which display the cards for each news article -->
+
 <!DOCTYPE html>
 <html lang="en">
   <head>
