+++ playerCard.html
@@ -1,54 +1,64 @@
+<!-- Page which displays the cats for a player to choose from -->
 <!DOCTYPE html>
-<!-- Coding By CodingNepal - codingnepalweb.com -->
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Popup Modal Box</title>
+    <title>Popup Box</title>
     <!-- CSS -->
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM"
+      crossorigin="anonymous"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css"
+    />
     <link rel="stylesheet" href="./css/playerCard.css" />
-    <!-- Fontawesome CDN Link -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css"/>
   </head>
   <body>
-    <section>
-      <div class="card" style="width: 18rem;">
-        <img src="..." class="card-img-top" alt="...">
-        <div class="card-body">
-          <h5 class="card-title">Card title</h5>
-          <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-          <a href="#" class="btn btn-primary">Go somewhere</a>
+    <div class="card-container">
+      <header>
+        <h1>Choose A Character</h1>
+      </header>
+      <div class="cards">
+        <div class="card" style="width: 18rem">
+          <img
+            src="./images/Player1/run/skeleton-04_running_00.png"
+            class="card-img-top"
+            alt="..."
+          />
+          <div class="card-body">
+            <h5 class="card-title">Pascal</h5>
+            <p class="card-text">
+              Not all heroes wear capes. Some have collars.
+            </p>
+            <button type="button" id="selectPlayer1" class="btn btn-primary">
+              Pick Me
+            </button>
+          </div>
+        </div>
+
+        <div class="card" style="width: 18rem">
+          <img
+            src="./images/Player2/run/skeleton-002_run_00.png"
+            class="card-img-top"
+            alt="..."
+          />
+          <div class="card-body">
+            <h5 class="card-title">Banjo-Kazooie</h5>
+            <p class="card-text">I came, I purred, I conquered.</p>
+            <button type="button" id="selectPlayer2" class="btn btn-primary">
+              Pick Me
+            </button>
+          </div>
+        </div>
       </div>
     </div>
 
-
-      <!-- <button class="show-modal">Show Modal</button>
-      <span class="overlay"></span>
-      <div class="modal-box">
-        <i src="./images/"></i>
-        <h2>Completed</h2>
-        <h3>You have sucessfully downloaded all the source code files.</h3>
-        <div class="buttons">
-          <button class="close-btn">Ok, Close</button>
-          <button>Open File</button>
-        </div>
-      </div> -->
-    </section>
-    <script>
-      const section = document.querySelector("section"),
-        overlay = document.querySelector(".overlay"),
-        showBtn = document.querySelector(".show-modal"),
-        closeBtn = document.querySelector(".close-btn");
-      
-        showBtn.addEventListener("click", () => section.classList.add("active"));
-
-        overlay.addEventListener("click", () =>
-          section.classList.remove("active"));
-
-        closeBtn.addEventListener("click", () =>
-          section.classList.remove("active"));
-
-    </script>
+    <script src="./js/playerCard.js"></script>
   </body>
 </html>
