--- playerStates.js
@@ -1,192 +1,0 @@
-import { Dust, Fire, Splash } from "./particles.js";
-
-const states = {
-  SITTING: 0,
-  RUNNING: 1,
-  JUMPING: 2,
-  FALLING: 3,
-  ROLLING: 4,
-  DIVING: 5,
-  HIT: 6,
-};
-
-class State {
-  constructor(state, game) {
-    this.state = state;
-    this.game = game;
-  }
-}
-
-export class Sitting extends State {
-  constructor(game) {
-    super("SITTING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;  //coordinates in the sheet where image starts X axis
-    this.game.player.frameY = 5;  //coordinates in the sheet where image starts Y axis
-    this.game.player.maxFrame = 4;  //number of animation frames in that sprite sheet
-  }
-  handleInput(input) {
-    if (input.includes("ArrowLeft") || input.includes("ArrowRight")) {
-      this.game.player.setState(states.RUNNING, 1); //the 1 here includes the movement speed of background
-    } else if (input.includes("ArrowUp")) {
-      this.game.player.setState(states.JUMPING, 1); 
-    } else if (input.includes("Control")) {
-      this.game.player.setState(states.ROLLING, 2); // the 2 is the movement speed of background
-    }
-  }
-}
-
-export class Running extends State {
-  constructor(game) {
-    super("RUNNING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;  //coordinates of frame in sprite sheet X axis
-    this.game.player.frameY = 3;  //coordinates of frame in sprite sheet Y axis
-    this.game.player.maxFrame = 8;  // number of frames for this animation
-  }
-  handleInput(input) {
-    if (this.game.player.speed > 0) {
-      this.game.particles.unshift( //adds one or more elements to the beginning of an array
-        new Dust(                   //and returns the new length of the array
-          this.game,                //adds the new particle and deletes the old one
-          this.game.player.x + this.game.player.width * 0.6,
-          this.game.player.y + this.game.player.height - 5
-        )
-      );
-    }
-    if (input.includes("ArrowDown")) {
-      this.game.player.setState(states.SITTING, 0);
-    } else if (input.includes("ArrowUp")) {
-      this.game.player.setState(states.JUMPING, 1);
-    } else if (input.includes("Control")) {
-      this.game.player.setState(states.ROLLING, 2);
-    }
-  }
-}
-
-export class Jumping extends State {
-  constructor(game) {
-    super("JUMPING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;
-    if (this.game.player.onGround()) this.game.player.vy -= 26;
-    this.game.player.frameY = 1;
-    this.game.player.maxFrame = 6;
-  }
-  handleInput(input) {
-    if (this.game.player.vy > this.game.player.weight) {
-      this.game.player.setState(states.FALLING, 1);
-    } else if (input.includes("Control")) {
-      this.game.player.setState(states.ROLLING, 2);
-    } else if (input.includes("ArrowDown")) {
-      this.game.player.setState(states.DIVING, 0);
-    }
-  }
-}
-
-export class Falling extends State {
-  constructor(game) {
-    super("FALLING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;
-    this.game.player.frameY = 2;
-    this.game.player.maxFrame = 6;
-  }
-  handleInput(input) {
-    if (this.game.player.onGround()) {
-      this.game.player.setState(states.RUNNING, 1);
-    } else if (input.includes("ArrowDown")) {
-      this.game.player.setState(states.DIVING, 0);
-    }
-  }
-}
-
-export class Rolling extends State {
-  constructor(game) {
-    super("ROLLING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;
-    this.game.player.frameY = 6;
-    this.game.player.maxFrame = 6;
-  }
-  handleInput(input) {
-    this.game.particles.unshift(  //unshiting / eliminating particles for FIRE
-      new Fire(
-        this.game,
-        this.game.player.x + this.game.player.width * 0.5,
-        this.game.player.y + this.game.player.height * 0.5
-      )
-    );
-    if (!input.includes("Control") && this.game.player.onGround()) {
-      this.game.player.setState(states.RUNNING, 1); //defaults ro running after movement
-    } else if (!input.includes("Control") && !this.game.player.onGround()) {
-      this.game.player.setState(states.FALLING, 1);
-    } else if (
-      input.includes("Control") &&
-      input.includes("ArrowUp") &&
-      this.game.player.onGround()
-    ) {
-      this.game.player.vy -= 27;
-    } else if (input.includes("ArrowDown") && !this.game.player.onGround()) {
-      this.game.player.setState(states.DIVING, 0);
-    }
-  }
-}
-
-export class Diving extends State {
-  constructor(game) {
-    super("DIVING", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;
-    this.game.player.frameY = 6;
-    this.game.player.maxFrame = 6;
-    this.game.player.vy = 20;       // makes the down velocity very fast when diving down
-  }
-  handleInput(input) {
-    this.game.particles.unshift(
-      new Fire(
-        this.game,
-        this.game.player.x + this.game.player.width * 0.5,
-        this.game.player.y + this.game.player.height * 0.5
-      )
-    );
-    if (this.game.player.onGround()) {
-      this.game.player.setState(states.RUNNING, 1);
-      for (let i = 0; i < 30; i++) {
-        this.game.particles.unshift(
-          new Splash(
-            this.game,
-            this.game.player.x + this.game.player.width * 0.5,
-            this.game.player.y + this.game.player.height * 0.5
-          )
-        );
-      }
-    } else if (input.includes("Control") && this.game.player.onGround()) {
-      this.game.player.setState(states.ROLLING, 2);
-    }
-  }
-}
-
-export class Hit extends State {
-  constructor(game) {
-    super("HIT", game);
-  }
-  enter() {
-    this.game.player.frameX = 0;
-    this.game.player.frameY = 4;
-    this.game.player.maxFrame = 10;
-  }
-  handleInput(input) {
-    if (this.game.player.frameX >= 10 && this.game.player.onGround()) {
-      this.game.player.setState(states.RUNNING, 1);
-    } else if (this.game.player.frameX >= 10 && !this.game.player.onGround()) {
-      this.game.player.setState(states.FALLING, 1);
-    }
-  }
-}
